vlm_decision_lab/requirements.txt
torch
transformers
torchvision
pillow
matplotlib
pyyaml
numpy
opencv-python
pandas
# Retrieval / similarity
# CLIP already via transformers
scikit-learn
# Optional acceleration
accelerate
# Progress + utilities
tqdm

